+++ apps/web/src/components/scenario-form/vatsim-import-dialog.tsx
@@ -1,3 +1,4 @@
+import { fetchPlanByCallsign } from "@/api/vatsim/fetch-plan-by-callsign";
 import { Input } from "@/components/ui/input";
 import { getRandomAltimeter } from "@workspace/plantools";
 import { AlertTriangleIcon, ImportIcon, Loader2 } from "lucide-react";
@@ -14,7 +15,6 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "../ui/dialog";
-import { fetchPlanByCallsign } from "./actions";
 
 export function VatsimImportDialog() {
   const [callsign, setCallsign] = useState("");
